requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2
django-cors-headers==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.10.0
psycopg2==2.7.7
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
